app/src/main/java/id/andiwijaya/hogwarts/data/local/dao/RemoteKeysDao.kt
package id.andiwijaya.hogwarts.data.local.dao

import androidx.room.Dao
import androidx.room.Insert
import androidx.room.OnConflictStrategy
import androidx.room.Query
import id.andiwijaya.hogwarts.domain.model.RemoteKeys

@Dao
interface RemoteKeysDao {
    @Insert(onConflict = OnConflictStrategy.REPLACE)
    suspend fun insertAll(remoteKey: List<RemoteKeys>)

    @Query("SELECT * FROM remote_keys WHERE id = :id")
    suspend fun getRemoteKeysId(id: String): RemoteKeys?

    @Query("SELECT COUNT(*) FROM remote_keys")
    suspend fun getNumberOfRecord(): Int
}